Reduce the redundancy of multiple extractiosn of the index file if ti is already present

+++ tool/bwa_aligner.py
@@ -98,16 +98,19 @@ class bwaAlignerTool(Tool):
         g_dir = genome_idx.split("/")
         g_dir = "/".join(g_dir[:-1])
 
-        try:
-            tar = tarfile.open(genome_idx)
-            tar.extractall(path=g_dir)
-            tar.close()
-        except IOError:
-            return False
+        if os.path.isfolder(g_dir) is False:
+            try:
+                tar = tarfile.open(genome_idx)
+                tar.extractall(path=g_dir)
+                tar.close()
+            except IOError:
+                return False
 
         gfl = genome_file_loc.split("/")
         genome_fa_ln = genome_idx.replace('.tar.gz', '/') + gfl[-1]
-        shutil.copy(genome_file_loc, genome_fa_ln)
+
+        if os.path.isFile(genome_fa_ln) is False:
+            shutil.copy(genome_file_loc, genome_fa_ln)
 
         if (
                 os.path.isfile(genome_fa_ln) is False or
